// src/Models/IPerson.ts
export interface IPerson{
   Title?: string;
   DisplayName?: string;
   AccountName: string;
   FirstName?: string;
   LastName?: string;
   PreferredName?: string;
   Email?: string;
   Manager?: string;
   AboutMe?: string;
   UserName?: string;
   DirectReports?: string[];
   PictureUrl?: string;
   ExtendedManagers?: string[];
   UserUrl?: string;
   MobilePhone?: string;
}